Improve Linux support
