+++ Cargo.toml
@@ -19,7 +19,6 @@ base64 = "0.13.0"
 block-modes="0.8.1"
 byteorder = "1.3.4"
 hex = "0.4.2"
-hidapi = "1.2.3"
 num = "0.4.0"
 pad = "0.1.6"
 ring = "0.16.20"
@@ -28,3 +27,8 @@ serde_cbor = "0.11.1"
 strum = "0.24"
 strum_macros = "0.24"
 x509-parser = "0.13"
+
+[dependencies.hidapi]
+version = "1.2.3"
+default-features = false
+features=["linux-static-hidraw"]
